push

+++ src/main/java/com/app/oc/service/MemberService.java
@@ -1,19 +1,20 @@
 package com.app.oc.service;
 
-import com.app.oc.dto.mypage.MemberDto;
+import com.app.oc.dto.mypage.MemberRequestDto;
+import com.app.oc.dto.mypage.MemberResponseDto;
+import com.app.oc.dto.mypage.PwdDto;
 import com.app.oc.dto.mypage.ResponseMemberDto;
-import com.app.oc.entity.Address;
-import com.app.oc.entity.AttenItem;
-import com.app.oc.entity.AttenShop;
-import com.app.oc.entity.Member;
+import com.app.oc.entity.*;
 import com.app.oc.repository.AttenShopRepository;
 import com.app.oc.repository.MemberRepository;
+import com.app.oc.repository.ShopRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 //기능을 정의할 수 있고 ,트잭션 관리
 @Service
@@ -23,13 +24,40 @@ import java.util.List;
 public class MemberService {
 
     private final MemberRepository memberRepository;
+    private final ShopRepository shopRepository;
     private final AttenShopRepository attenShopRepository;
 
+
+    /**
+     * 회원가입
+     * @param memberRequestDTO
+     * @return
+     */
+    @Transactional
+    public MemberResponseDto signup(MemberRequestDto memberRequestDTO) {
+
+        Member newMember = new Member(memberRequestDTO);
+        validateDuplicateMember(newMember); //중복회원 검증
+
+        return MemberResponseDto.of(memberRepository.saveAndFlush(newMember));
+    }
+
+    /**
+     * 중복 회원 검증
+     * @param member
+     */
+    public void validateDuplicateMember(Member member) {
+        Optional<Member> findMember = memberRepository.findById(member.getMemberId());
+        if(!findMember.isEmpty()) {
+            throw new IllegalStateException("이미 존재하는 회원입니다.");
+        }
+    }
+
     //Member 1명 찾기
-    @Transactional(readOnly=true)// jpa 변경감지 내부기능 활성화x,update시 정합성 유지,
-    public Member findOne(String id) {
-        return memberRepository.findById(id)
-                .orElseThrow(()->new IllegalArgumentException("ID가 없습니다."));
+    @Transactional(readOnly=true)
+    public Member findOne(String memberId) {
+        return memberRepository.findById(memberId)
+                .orElseThrow(() -> new IllegalArgumentException("ID가 없습니다."));
     }
 
 
@@ -41,21 +69,40 @@ public class MemberService {
 
         Address address = new Address(buyer.getPostcode(),buyer.getAddress1(),buyer.getAddress2());
         Member update = buyerEntity.update(buyer,address);
-        log.info("update : {}",update);
         return buyerEntity.getMemberId();
     }
 
-    //Member 삭제
-    public String delete(String id) {
+    //Member 삭제 -item 삭제
+    public String delete(String id, String pwd) {
+
         Member member = memberRepository.findById(id).orElseThrow(() -> new IllegalArgumentException("ID가 없습니다."));
+        if (!member.getPassword().equals(pwd)) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+
+        if (member.getRole() == MemberRole.SELLER) {
+            System.out.println("member = " + member);
+
+            List<ShoppingMal> shoppingMals = member.getShoppingMals();
+            System.out.println("shoppingMals = " + shoppingMals);
+            for (ShoppingMal shoppingMal : shoppingMals) {
+                 shopRepository.delete(shoppingMal);
+                System.out.println("shoppingMal = " + shoppingMal);
+
+            }
+        }
         memberRepository.delete(member);
         return "ok"; 
     }
 
     //비밀번호 update
-    public String updatePwd(String id, String nPwd) {
+    public String updatePwd(String id, PwdDto pwdDto ) {
     Member member = memberRepository.findById(id).orElseThrow(()->new IllegalArgumentException("ID가 없습니다"));
-    member.updatePwd(nPwd);
+    if (!member.getPassword().equals(pwdDto.getPwd())) {
+            throw new IllegalArgumentException("비밀번호가 일치하지 않습니다.");
+        }
+
+    member.updatePwd(pwdDto.getN_pwd());
     return member.getMemberId();
 }
 
@@ -72,4 +119,7 @@ public class MemberService {
     public List<AttenItem> findByAttenItem(String id) {
         return attenShopRepository.findAttenItem(id);
     }
+
+
+
 }
